Move table/bloom_block.* -> table/plain/plain_table_bloom.*
Summary: table/bloom_block (BloomBlockBuilder) is only used by the plain
table, so moving it there. This sets up more refactoring of legacy
Bloom filter implementations.

Test Plan: make check
