+++ TARGETS
@@ -229,7 +229,6 @@ cpp_library(
         "table/block_based/partitioned_filter_block.cc",
         "table/block_based/uncompression_dict_reader.cc",
         "table/block_fetcher.cc",
-        "table/bloom_block.cc",
         "table/cuckoo/cuckoo_table_builder.cc",
         "table/cuckoo/cuckoo_table_factory.cc",
         "table/cuckoo/cuckoo_table_reader.cc",
@@ -239,6 +238,7 @@ cpp_library(
         "table/merging_iterator.cc",
         "table/meta_blocks.cc",
         "table/persistent_cache_helper.cc",
+        "table/plain/plain_table_bloom.cc",
         "table/plain/plain_table_builder.cc",
         "table/plain/plain_table_factory.cc",
         "table/plain/plain_table_index.cc",
