+++ table/plain/plain_table_bloom.cc
@@ -3,7 +3,7 @@
 //  COPYING file in the root directory) and Apache 2.0 License
 //  (found in the LICENSE.Apache file in the root directory).
 
-#include "table/bloom_block.h"
+#include "table/plain/plain_table_bloom.h"
 
 #include <string>
 #include "rocksdb/slice.h"
