+++ table/plain/plain_table_builder.cc
@@ -19,9 +19,9 @@
 #include "rocksdb/options.h"
 #include "rocksdb/table.h"
 #include "table/block_based/block_builder.h"
-#include "table/bloom_block.h"
 #include "table/format.h"
 #include "table/meta_blocks.h"
+#include "table/plain/plain_table_bloom.h"
 #include "table/plain/plain_table_factory.h"
 #include "table/plain/plain_table_index.h"
 #include "util/coding.h"
